+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### W.I.P
+
+  - Fixed sticky wrapper resize on window resize/orientation change
+
+
 ### 2.17.0 (February 20, 2015)
 
   - IMPORTANT: Removed caption component (Use overlay component)